[MM] no error on run initial db creation

+++ database/migrate.sh
@@ -1,11 +1,11 @@
 #!/bin/sh
 
 # psql -f $FILE -d fridge
-# if [ -f "fridge.db" ]; then
 
-
-echo "Removing old database..."
-rm fridge.db
+if [ -f "fridge.db" ]; then
+    echo "Removing old database..."
+    rm fridge.db
+fi
 
 echo "Applying schema..."
 FILES=$(ls database/migration/*.sql)
